requirements.txt
Django==2.2.16
django-environ==0.4.5
mysqlclient==1.4.6
Jinja2==2.11.2
django-debug-toolbar==3.1.1
django-js-reverse==0.9.1
django-decorator-include==3.0
python-memcached==1.59
django-cache-machine==1.1.0
boto==2.49.0
djangorestframework==3.12.0
djangorestframework-camel-case==1.2.0
drf-writable-nested==0.6.1
cached-property==1.5.2
phonenumbers==8.12.9
django-phonenumber-field==4.0.0
django-filter==2.4.0
django-webpack-loader==0.7.0
requests==2.24.0
lxml==4.5.2
Pillow==7.2.0
